Make code follow RESTful concept
