+++ portfolio/src/main/webapp/leave-comment/leave-comment.html
@@ -14,7 +14,7 @@
   <body>
     <div id="leave-comment-form">
       Please share your feedback!
-      <form method="post" action="../add-comment">
+      <form method="post" action="../comments">
         <input name="email" type="email" placeholder="Your email adress">
         <textarea name="content" wrap="soft" placeholder="Your comment"></textarea>
         <input class="button" type="submit" placeholder="Submit!">
